+++ CLIBuilder/CmdTree/clistd.cpp
@@ -166,7 +166,7 @@ clistd_validate_leaf (tlv_struct_t *tlv) {
 /* Standard Validations End */
 
 int
-clistd_config_device_default_handler (int cmdcode,  Stack_t *tlv_stack, op_mode enable_or_disable) {
+clistd_config_device_default_handler (int64_t cmdcode,  Stack_t *tlv_stack, op_mode enable_or_disable) {
 
     tlv_struct_t *tlv = NULL;
     
@@ -185,7 +185,7 @@ clistd_config_device_default_handler (int cmdcode,  Stack_t *tlv_stack, op_mode 
 }
 
 int
-show_help_handler(int cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable){
+show_help_handler(int64_t cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable){
 
     attron(COLOR_PAIR(GREEN_ON_BLACK));
     printw("\nWelcome to Help Wizard\n");
@@ -217,14 +217,14 @@ show_help_handler(int cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable){
 }
 
 int
-show_history_handler (int cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable) {
+show_history_handler (int64_t cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable) {
 
     cli_history_show ();
     return 0;
 }
 
 int
-cli_terminate_handler(int cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable){
+cli_terminate_handler(int64_t cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable){
 
     endwin();
     printw("Bye Bye\n");
