+++ Layer5/app_handlers.h
@@ -6,7 +6,7 @@
 typedef struct stack Stack_t ;
 
 int
-spf_algo_handler(int cmdcode, Stack_t *tlv_stack,
+spf_algo_handler(int64_t cmdcode, Stack_t *tlv_stack,
                           op_mode enable_or_disable);
 
 int
@@ -56,6 +56,6 @@ param_t *
 vrf_build_config_tree (param_t *node_name);
 
 int
-ping_handler(int cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable);
+ping_handler(int64_t cmdcode, Stack_t *tlv_stack, op_mode enable_or_disable);
 
 #endif /* __APP_HANDLERS__ */
